NXP-24034: Extract current trash service to LifeCycleTrashService
