+++ nuxeo-core/nuxeo-core/pom.xml
@@ -29,6 +29,10 @@
       <artifactId>nuxeo-runtime-datasource</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.nuxeo.runtime</groupId>
+      <artifactId>nuxeo-runtime-migration</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.nuxeo.ecm.core</groupId>
       <artifactId>nuxeo-core-api</artifactId>
     </dependency>
